one_running_test/requirements.txt
Django==5.2
mysqlclient==2.2.7
celery==5.3.1
redis==4.5.5
djangorestframework==3.14.0
django-celery-beat==2.8.0